+++ Sources/BraintreePayPal/BTPayPalClient.swift
@@ -108,7 +108,7 @@ import BraintreeDataCollector
     @objc(initWithAuthorization:universalLink:fallbackURLScheme:)
     public convenience init(authorization: String, universalLink: URL, fallbackURLScheme: String? = nil) {
         self.init(authorization: authorization)
-        
+
         /// appending a PayPal app switch specific path to verify we are in the correct flow when
         /// `canHandleReturnURL` is called
         self.universalLink = universalLink.appendingPathComponent("braintreeAppSwitchPayPal")